SCSA_Project_CRUD-main/user_sign_up/urls.py
from django.urls import path
from .views import UserView, UserRegisterView
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)


urlpatterns = [
    path("", UserView.as_view(), name="user_info"),
    path("register/", UserRegisterView.as_view(), name="user_register"),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
