SCSA_Project_CRUD-main/requirements.txt
django>=4.0.2
djangorestframework
pyyaml
requests
django-cors-headers